fix: User service works with User (not user request)
